Added DimensionsConfiguration
works! Probably some quirks to fix though
