+++ src/main/java/com/code12/worldtp/gui/settings/SelectWorldToEditGui.java
@@ -1,0 +1,57 @@
+package com.code12.worldtp.gui.settings;
+
+import com.code12.worldtp.files.DataManager;
+import com.code12.worldtp.files.References;
+import com.github.stefvanschie.inventoryframework.gui.GuiItem;
+import com.github.stefvanschie.inventoryframework.gui.type.ChestGui;
+import com.github.stefvanschie.inventoryframework.pane.StaticPane;
+import org.bukkit.Bukkit;
+import org.bukkit.entity.Player;
+import org.bukkit.inventory.ItemFlag;
+import org.bukkit.inventory.ItemStack;
+import org.bukkit.inventory.meta.ItemMeta;
+
+import java.util.List;
+
+public class SelectWorldToEditGui {
+    DataManager data = References.data;
+
+    public SelectWorldToEditGui(Player player){
+        ChestGui gui = new ChestGui(4, "Settings: Select World");
+        gui.setOnGlobalClick(event -> event.setCancelled(true));
+
+        StaticPane pane = new StaticPane(0, 0, 9, 4);
+
+        List<String> menuGroupList = data.getConfig().getStringList("menuGroupList");
+
+        int x = 0;
+        int y = 0;
+
+        for(String menuGroup : menuGroupList){
+            ItemStack item = data.getConfig().getItemStack("menuGroupID." + menuGroup + ".item");
+            String displayName = data.getConfig().getString("menuGroupID." + menuGroup + ".displayName");
+
+            ItemMeta itemMeta = item.getItemMeta();
+            itemMeta.setDisplayName(displayName);
+            itemMeta.addItemFlags(ItemFlag.HIDE_ATTRIBUTES);
+            item.setItemMeta(itemMeta);
+
+            GuiItem guiItem = new GuiItem(item, event ->{
+                new SettingsGui((Player) event.getWhoClicked(), Bukkit.getWorld(menuGroup));
+            });
+
+            pane.addItem(guiItem, x, y);
+
+            x++;
+            if(x >= 9){
+                x = 0;
+                y++;
+            }
+        }
+
+        gui.addPane(pane);
+        gui.show(player);
+    }
+
+
+}
